Rename filenames slave→manager
