+++ src/main/core/manager.c
@@ -14,7 +14,7 @@
 #include "main/core/master.h"
 #include "main/core/scheduler/scheduler.h"
 #include "main/core/scheduler/scheduler_policy.h"
-#include "main/core/slave.h"
+#include "main/core/manager.h"
 #include "main/core/support/definitions.h"
 #include "main/core/support/object_counter.h"
 #include "main/core/support/options.h"
