+++ src/main/core/manager.h
